// ButorRevolutionWebAPI/src/frontend/web/projects/factory/src/app/admin/users/models/users.models.ts
export interface IUserEditViewModel {
    basicInfo?: IUserBasicInfoViewModel;
    isEmployee?: boolean;
    isActivated?: boolean;
    claims?: number[];
    daysoff?: ICalendarViewModel;
    workingInfo?: IUserWorkingInfoViewModel;
    notifications?: IUserNotificationDataModel;

}

export interface IUserBasicInfoViewModel {
    id?: number;
    name: string;
    roles?: number[];
    status?: string;
    company?: number;
    companyList?: ICompanySelectList;
    address?: IAddressViewModel;
    email?: string;
    phone?: string;
    language?: LanguageTypeEnum;
    addedOn?: string;
    image?: ImageDetailsViewModel;
}
export interface ImageDetailsViewModel {
    containerName?: string;
    fileName?: string;
}
export interface ICompanySelectList {
    id: number;
    name?: string;
}
export interface ILanguageSelectList {
    id: number;
    name?: string;
}

export interface IAddressViewModel {
    address?: string;
    postCode?: number;
    city?: string;
    countryId?: number;
}

export interface IUserWorkingInfoViewModel {
    usersOffices?: number[];
    usersOfficesList?: UsersOfficesModel[];
    supervisors?: number;
    supervisorsList?: IUserRole;
    supervisees?: IUserRole[];
    groups?: string[];
    discount?: { from: number, to: number };
    workingHours?: IUserWorkingHoursModel[];
}

export interface IUserWorkingHoursModel {
    day?: string;
    isChecked?: boolean;
    dayTypeId?: number;
    workingHour?: IUserWorkingHour[];
}

export interface IUserWorkingHour {
    dayTypeId: number;
    id: number;
    from: string | Date;
    to: string | Date;
}

export interface IModuleViewModel {
    name: string;
    description: string;
    division?: DivisionTypeEnum;
    claims: IClaimViewModel[];
}

export interface IClaimViewModel {
    id: number;
    name: string;
    division?: DivisionTypeEnum;
    isChecked?: boolean;
}

export interface ICalendarViewModel {
    sickLeave: AbsenceDayViewModel[];
    daysOff: AbsenceDayViewModel[];
    deleted?: Date[];
}

export interface IUserRole {
    id: number;
    name?: string;
}

export interface IUserNotificationViewModel {
    label: string;
    notifications: IUserNotificationModel[];
}
export interface IUserNotificationDataModel {
    notificationType: string[];
    eventTypeIds: number[];
}
export interface IUserNotificationModel {
    id?: number;
    isChecked?: boolean;
    name: string;
}

export interface AbsenceDayViewModel {
    id?: number;
    date: Date;
    absenceType?: AbsenceTypeEnum;
}

export interface UserSalesModel {
    id: number;
    name?: string;
    phoneNumber?: string;
    email?: string;
}

export interface CompanySelectModel {
    id: number;
    name: string;
}
export interface RolesSelectModel {
    id: number;
    name: string;
}
export interface CountryListSelectModel {
    id: number;
    name: string;
}
export interface UsersOfficesModel {
    id: number;
    name: string;
}

export interface DayTypeListModel {
    id: number;
    dayType: DayTypeEnum;
    translation: string;
    order: number;
}